requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
idna==3.6
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
